adapters/python/ollama_adapter/requirements.txt
grpcio==1.64.0
grpcio-tools==1.64.0
protobuf==5.27.2
uvloop==0.20.0
orjson==3.10.7
fastapi==0.115.0
uvicorn==0.30.6
